Handle -Wconversion compile flags
To support more compiler as ms one, add more constraints
on warning generation. ie. add -Wconversion flags on gcc.

Also fix build issues trig by this activation.

+++ parameter/BitParameterType.cpp
@@ -184,7 +184,7 @@ bool CBitParameterType::toBlackboard(uint64_t uiUserValue, uint64_t& uiValue, CP
 
 void CBitParameterType::fromBlackboard(uint32_t& uiUserValue, uint64_t uiValue, CParameterAccessContext& /*ctx*/) const
 {
-    uiUserValue = (uiValue & getMask()) >> _bitPos;
+    uiUserValue = uint32_t((uiValue & getMask()) >> _bitPos);
 }
 
 // Access from area configuration
